Generalize "clearing the unused cache keys" example
This change makes the concurrency issue handled in the cleanup handler
of the first weakref + finalizationgroup example more prominent.  It
also explicitly makes the connection with weak-value maps.

Some of the changes are quite subjective, so review either way is
welcome.  I could make a smaller change if that's more appropriate; I'm
still learning the culture here :)